[CHORE] 운동 기록 뷰 이동 #41
